`fs-promise` util does not modify `fs` standard library
`fs-promise` was altering `fs` by adding `-Async` methods. In result
it was impossible to deploy with Claudia any project which uses
`Promise.promisifyAll(fs)` (or if any of its dependency does so)
because that operation want to add `-Async` too.

In this commit `fs-promise` no longer alters `fs` and, instead,
returns just a couple of "asynced" methods.